refactor(types): share `module-scan` API types
This moves types needed for the `/scan/status` endpoint into `@votingworks/types`. It makes them accessible from a "vanity" path of `@votingworks/types/api/module-scan` by putting some alias files at `api/module-scan` within `@votingworks/types` that point to the real ones.
