+++ apps/module-scan/tsconfig.test.json
@@ -1,5 +1,8 @@
 {
   "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
   "include": ["src", "test"],
   "exclude": []
 }
