+++ libs/types/.eslintignore
@@ -1,1 +1,2 @@
-jest.config.js
+*.js
+*.d.ts
