+++ libs/types/api/index.d.ts
@@ -1,0 +1,1 @@
+export * from '../dist/src/api'
