+++ libs/types/api/module-scan/index.js
@@ -1,0 +1,1 @@
+module.exports = require('../../dist/src/api/module-scan')
