+++ libs/types/src/api/index.ts
@@ -1,0 +1,1 @@
+export type ISO8601Timestamp = string
